Refactor web composition root
